# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OpenAI-совместимый HTTP API (chat, embeddings, models) и RAG поверх PostgreSQL/pgvector.",
        "title": "llama-cpp-gpt-api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v1/chat/completions": {
            "post": {
                "description": "OpenAI-совместимая генерация ответа. При включённом RAG в system подмешиваются фрагменты из базы.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Chat completions",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ReqChatCompletion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ResChatCompletion"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    }
                }
            }
        },
        "/v1/embeddings": {
            "get": {
                "description": "Векторное представление текста (POST JSON или GET query).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "embeddings"
                ],
                "summary": "Embeddings",
                "parameters": [
                    {
                        "description": "Тело (POST)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.ReqEmbeddings"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Модель",
                        "name": "model",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Текст",
                        "name": "input",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ResEmbeddings"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    }
                }
            },
            "post": {
                "description": "Векторное представление текста (POST JSON или GET query).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "embeddings"
                ],
                "summary": "Embeddings",
                "parameters": [
                    {
                        "description": "Тело (POST)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/types.ReqEmbeddings"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Модель",
                        "name": "model",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Текст",
                        "name": "input",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ResEmbeddings"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    }
                }
            }
        },
        "/v1/models": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Список моделей",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ResModelsList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    }
                }
            }
        },
        "/v1/models/{model}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Модель по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID модели",
                        "name": "model",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ModelObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.Choice": {
            "type": "object",
            "properties": {
                "delta": {
                    "$ref": "#/definitions/types.Message"
                },
                "finish_reason": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "message": {
                    "$ref": "#/definitions/types.Message"
                }
            }
        },
        "types.Embedding": {
            "type": "object",
            "properties": {
                "embedding": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "index": {
                    "type": "integer"
                }
            }
        },
        "types.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "types.ModelObject": {
            "type": "object",
            "properties": {
                "capabilities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "owned_by": {
                    "type": "string"
                }
            }
        },
        "types.OpenAIErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/types.OpenAIErrorDetail"
                }
            }
        },
        "types.OpenAIErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "param": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "types.RAGFragment": {
            "type": "object",
            "properties": {
                "chunk_index": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                },
                "source_path": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "types.RAGInfo": {
            "type": "object",
            "properties": {
                "chunks_used": {
                    "type": "integer"
                },
                "fragments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.RAGFragment"
                    }
                },
                "titles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "types.ReqChatCompletion": {
            "type": "object",
            "properties": {
                "max_tokens": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Message"
                    }
                },
                "model": {
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                },
                "rag_top_k": {
                    "type": "integer"
                },
                "seed": {
                    "type": "integer"
                },
                "stream": {
                    "type": "boolean"
                },
                "temperature": {
                    "type": "number"
                },
                "top_p": {
                    "type": "number"
                },
                "use_rag": {
                    "description": "nil → из конфига RAG.UseInChatCompletions (по умолчанию true)",
                    "type": "boolean"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "types.ReqEmbeddings": {
            "type": "object",
            "required": [
                "input"
            ],
            "properties": {
                "input": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
        "types.ResChatCompletion": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Choice"
                    }
                },
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "rag": {
                    "$ref": "#/definitions/types.RAGInfo"
                },
                "usage": {
                    "$ref": "#/definitions/types.Usage"
                }
            }
        },
        "types.ResEmbeddings": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Embedding"
                    }
                },
                "model": {
                    "type": "string"
                }
            }
        },
        "types.ResModelsList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ModelObject"
                    }
                },
                "object": {
                    "type": "string"
                }
            }
        },
...
        "types.Usage": {
            "type": "object",
            "properties": {
                "completion_tokens": {
                    "type": "integer"
                },
                "prompt_tokens": {
                    "type": "integer"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        }
    }
}